split rep_log_ets into snapshot, backfill, common and normal ops
backfill - CRUD operations on new relation not previously part of
logical replication will result in a push callback to this module
fetching metadata and then returning to normal operations.

snapshot - at the start of the replication process we push
callback to this module to fetch the initial replication snapshot
with the transaction snapshot.

common - some common functions used by both backfill and shapshot.

pgmp_rep_log_ets - normal CRUD replication operations are processed in
this module. Transitions to snapshot (once only) at the start of
replication, and to backfill only when a new relation is processed by
relication.